requirements.txt
streamlit==1.38.0
tensorflow==2.17.0
pillow==10.4.0
numpy
opencv-python